Create a separate google sign in and email log in screens. Fixed issue with selecting the back button on going back to the login screen.
