+++ lib/splash_screen.dart
@@ -1,15 +1,8 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'dart:async';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_widget_app/authentication.dart';
 import 'package:flutter_auth_buttons/flutter_auth_buttons.dart';
 import 'package:flutter_widget_app/global_contants.dart';
-
-class LoginData {
-  String email = '';
-  String password = '';
-}
 
 class SplashPage extends StatefulWidget {
   @override
@@ -17,178 +10,19 @@ class SplashPage extends StatefulWidget {
 }
 
 class _SplashPageState extends State<SplashPage> {
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-  final GlobalKey<FormState> _formKey = new GlobalKey<FormState>();
-  LoginData _data = new LoginData();
   bool _isLoading = false;
 
-  String validateEmail(String value) {
-    if (value.isEmpty || !value.contains('@')) {
-      return 'The E-mail Address must be a valid email address.';
-    }
-    return null;
-  }
-
-  String validatePassword(String value) {
-    if (value.length < 6) {
-      return 'The Password must be at least 6 characters.';
-    }
-    return null;
-  }
-
-  Future signInWithEmail() async {
-    FirebaseUser user;
-
-    // First validate form.
-    if (this._formKey.currentState.validate()) {
-       _formKey.currentState.save(); // Save our form now.
-
-       setState(() {
-         _isLoading = true;
-       });
-
-      try {
-        user = await _auth.signInWithEmailAndPassword(email: _data.email,
-            password: _data.password);
-      } catch (error) {
-        showErrorDialog(error);
-      } finally {
-        assert(user != null);
-        assert(await user.getIdToken() != null);
-
-        final FirebaseUser currentUser = await _auth.currentUser();
-        assert(user.uid == currentUser.uid);
-
-        print('signInEmail succeeded');
-        setState(() {
-          _isLoading = false;
-        });
-
-        // Navigate to main calendar view
-        _navigateToCalendarView();
-      }
-    }
-  }
-
   void _navigateToCalendarView() {
-    if (Navigator.canPop(context)) {
-      Navigator.of(context).pop();
-    } else {
-      Navigator.of(context).pushNamed(Constants.calendarRoute);
-    }
+    Navigator.of(context).pushNamedAndRemoveUntil(Constants.calendarRoute,
+            (Route<dynamic> route) => false);
   }
 
-  Future signUpWithEmail() async {
-    FirebaseUser user;
-
-    if (this._formKey.currentState.validate()) {
-      _formKey.currentState.save();
-
-      setState(() {
-        _isLoading = true;
-      });
-
-      try {
-        user = await _auth.createUserWithEmailAndPassword(
-            email: _data.email, password: _data.password);
-      } catch (error) {
-        showErrorDialog(error);
-      } finally {
-        assert(user != null);
-        assert(await user.getIdToken() != null);
-
-        final FirebaseUser currentUser = await _auth.currentUser();
-        assert(user.uid == currentUser.uid);
-
-        print('signInEmail succeeded');
-
-        // Add user to the 'users' collection
-        await Firestore.instance.collection('users').document()
-            .setData({'email': currentUser.email, 'token': ""});
-
-        setState(() {
-          _isLoading = false;
-        });
-
-        // Navigate to main calendar view
-        _navigateToCalendarView();
-      }
-    }
-  }
-
-  void showErrorDialog(error) {
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        return new AlertDialog(
-          title: new Text('Sign Up Error'),
-          content: new Text(error.message),
-          actions: <Widget>[
-            new FlatButton(
-              onPressed: () {
-                setState(() {
-                  _isLoading = false;
-                });
-                Navigator.of(context).pop(true);
-              },
-              child: new Text('OK')
-            )
-          ],
-        );
-      }
-    );
+  void _navigateToSignInPage() {
+    Navigator.of(context).pushNamed(Constants.signInRoute);
   }
 
   @override
   Widget build(BuildContext context) {
-    print('build().....');
-    final emailWidget = new TextFormField(
-      keyboardType: TextInputType.emailAddress,
-      decoration: new InputDecoration(
-          hintText: 'email@gmail.com',
-          labelText: 'Email address',
-          border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(32.0)
-          )
-      ),
-      style: TextStyle(fontSize: 20.0, color: Colors.black),
-      validator: this.validateEmail,
-      onSaved: (String value) {
-        this._data.email = value;
-      },
-    );
-
-    final passwordWidget = new TextFormField(
-      obscureText: true,
-      decoration: new InputDecoration(
-          hintText: 'Password',
-          labelText: 'Please enter your password',
-          border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(32.0)
-          )
-      ),
-      style: TextStyle(fontSize: 20.0, color: Colors.black),
-      validator: this.validatePassword,
-      onSaved: (String value) {
-        this._data.password = value;
-      },
-    );
-
-    final loginButton = new RaisedButton(
-        color: Colors.lightBlueAccent,
-        onPressed: () {
-          this.signInWithEmail();
-        },
-        child: new Text('Login', style: new TextStyle(fontSize: 24.0, color: Colors.white))
-    );
-
-    final signUpButton = new RaisedButton(
-      color: Colors.lightBlueAccent,
-      onPressed: this.signUpWithEmail,
-      child: new Text('Sign Up', style: new TextStyle(fontSize: 24.0, color: Colors.white))
-    );
-
     final signInWithGoogleButton = GoogleSignInButton(
       onPressed: () async {
         setState(() {
@@ -212,52 +46,26 @@ class _SplashPageState extends State<SplashPage> {
         if(user != null) {
           _navigateToCalendarView();
         } else {
-          print("Error signing with Google.");
+          print("Error signing in with Google.");
         }
       },
       darkMode: true);
 
-    final resetPasswordText = new GestureDetector(
-      onTap: () {
-        _formKey.currentState.save();
-        String resetMessage;
-        if (_data.email.isEmpty) {
-          resetMessage = 'Please enter a email address.';
-        } else {
-          resetMessage = 'Reset password for ' + _data.email;
-        }
-
-        showDialog(
-          context: context,
-          barrierDismissible: false,
-          builder: (BuildContext context) {
-            return new AlertDialog(
-              title: new Text('Reset Password'),
-              content: new Text(resetMessage),
-              actions: <Widget>[
-                new FlatButton(
-                    onPressed: () {
-                      if (_data.email.isNotEmpty) {
-                        _auth.sendPasswordResetEmail(email: _data.email);
-                      }
-                      Navigator.of(context).pop(true);
-                    },
-                    child: new Text('OK')
-                ),
-                new FlatButton(
-                    onPressed: () {
-                      Navigator.of(context).pop(true);
-                    },
-                    child: new Text("CANCEL"),
-                )
-              ],
-            );
-          }
-        );
-      },
-      child: new Text('Reset Password',textAlign: TextAlign.center,
-        style: new TextStyle(fontSize: 24.0, color: Colors.blue),
-      ),
+    final signInWithEmailButton = new Container(
+        child: MaterialButton(
+          onPressed: _navigateToSignInPage,
+          elevation: 4.0,
+          color: Color(0xFF4285F4),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Icon(Icons.email, color: Colors.white, size: 38),
+              SizedBox(width: 14.0,),
+              Text("Sign in with email",
+                  style: new TextStyle(fontSize: 18.0, color: Colors.white)),
+            ],
+          ),
+        )
     );
 
     final loginImage = new Image.asset('assets/calendar.png',
@@ -274,39 +82,19 @@ class _SplashPageState extends State<SplashPage> {
       backgroundColor: Colors.white,
       body: Padding(
         padding: const EdgeInsets.only(left: 24.0, right: 24.0),
-        child: Center(
+        child: SafeArea(
           child: Column(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              loginImage,
-              SizedBox(height: 16.0),
-              new Form(
-                key: this._formKey,
-                child: new Column(
-                  children: <Widget>[
-                    emailWidget,
-                    SizedBox(height: 8.0),
-                    passwordWidget,
-                  ],
-                ),
+              Center(child: Text("Events Calendar",
+                style: TextStyle(fontSize: 38, fontWeight: FontWeight.bold), textAlign: TextAlign.center)
               ),
-              SizedBox(height: 8.0),
-              _isLoading ? loadingSpinner :
-                  Column(
-                    children: <Widget>[
-                      SizedBox(height: 4.0),
-                      signInWithGoogleButton,
-                      Row(
-                        children: <Widget>[
-                          Expanded(child: loginButton),
-                          SizedBox(width: 4.0),
-                          Expanded(child:signUpButton),
-                        ],
-                      ),
-                    ],
-                  ),
+              loginImage,
+              signInWithGoogleButton,
+              Center(child: Text("or")),
               SizedBox(height: 4.0),
-              resetPasswordText,
+              signInWithEmailButton,
             ],
           ),
         ),
